# internal/docs/swagger.yaml
basePath: /api/v1/
definitions:
  controllers.ErrorResponse:
    properties:
      code:
        example: 400
        type: integer
      message:
        example: status bad request
        type: string
      more_info:
        example: http://
        type: string
    type: object
  models.Download:
    properties:
      audio_eta:
        type: string
      audio_ready:
        type: boolean
      completed:
        type: string
      done:
        type: boolean
      error:
        type: boolean
      filename:
        type: string
      running:
        type: boolean
      started:
        type: string
      std_err:
        items:
          type: string
        type: array
      std_out:
        items:
          type: string
        type: array
      url:
        type: string
      video_eta:
        type: string
      video_ready:
        type: boolean
    type: object
  models.Job:
    properties:
      download_info:
        $ref: '#/definitions/models.Download'
        type: object
      uuid:
        format: uuid
        type: string
    type: object
info:
  contact:
    email: support@egeback.se
    name: API Support
    url: http://xxxx.xxx.xx
  description: API to download with svt-download
  license:
    name: MIT License
    url: https://opensource.org/licenses/MIT
  title: Play Media API - Downloader
  version: "1.0"
paths:
  /jobs:
    get:
      consumes:
      - application/json
      description: List all jobs active in the API
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Job'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      summary: List jobs
      tags:
      - jobs
    post:
      consumes:
      - application/json
      description: Add job to API for download
      parameters:
      - description: url to download
        format: str
        in: query
        name: url
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Job'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      summary: Add job
      tags:
      - jobs
  /jobs/:uuid:
    get:
      consumes:
      - application/json
      description: Get a specific job by UUID
      parameters:
      - description: job uuid
        format: str
        in: path
        name: uuid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Job'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      summary: List jobs
      tags:
      - jobs
  /ping:
    get:
      consumes:
      - application/json
      description: do ping
      produces:
      - application/json
      responses:
        "200":
          description: pong
          schema:
            type: string
        "400":
          description: ok
          schema:
            type: string
        "404":
          description: ok
          schema:
            type: string
        "500":
          description: ok
          schema:
            type: string
      summary: ping common
      tags:
      - common
swagger: "2.0"